Add files via upload
The dataset comprises 1567 data points, each containing 591 features. These features represent various measurements associated with production entities, with corresponding pass/fail labels indicating the outcome of in-house line testing. A label of "-1" denotes a pass, while "1" indicates a fail, with each data point timestamped for the specific test.

PROJECT OBJECTIVE:
Our goal is to develop a classifier to predict the pass/fail yield of a given process entity. Additionally, we aim to analyze whether all features are necessary for building an effective model.